fix(UhyveIfVersion): assume big endian encoding
hermit-entry was modified against a local version of uhyve-interface
that did not have the big endian change, but this change was not sent
upstream to Uhyve as well. `from_ne_bytes`, which would use the host's
endianness (following the same logic of the rest of the big endian
support) is therefore wrong. Unfortunately, this was not caught before
Hermit's 0.12.0 release.

This change uses big endian for the Uhyve interface version number,
in line with the preexisting code in uhyve-interface.